decouple fetching logic from actions component
