+++ README.md
@@ -7,3 +7,9 @@
 ## Run
 
 `npm run dev`
+
+## Env variables
+
+Env variables are stored in [Vercel](https://vercel.com/glodollar/glo-wallet/settings/environment-variables).
+
+The easiest way to retrieve is to use [Vercel CLI](https://vercel.com/docs/cli) and run `vercel env pull`. Make sure you link your repo first with `vercel link`.
